chore: move templates api to app router
